requirements.txt
Django==4.0.4
psycopg2-binary==2.9.3

pip~=22.0.4
djangorestframework~=3.13.1
